R31c: fix 42 CI clippy warnings + auto-fix ~130 style warnings
- Fix empty_line_after_doc_comments in pager.rs, coverage_r7_push80b.rs
- Fix for_kv_map in cluster_mgmt.rs
- Add Default impl for 26+ structs (new_without_default)
- Fix match_like_matches_macro in dist_txn.rs
- Fix collapsible_if in dist_txn.rs, ext/advanced.rs
- Fix if_same_then_else in ha_dr.rs
- Fix manual_clamp in schema.rs
- Fix large_enum_variant: Box ConnStream::Tls in mysql.rs
- Fix redundant_closure in statement.rs
- Fix derivable_impls: WalSyncMode, GroupCommitConfig in wal.rs
- Add is_empty() to LruKEvictor, PlanCacheEvictor
- Fix map_clone in pager.rs
- Auto-fix ~130 additional warnings (single_match → if let, etc)
- All 5032 tests pass

+++ src/vm/execute_tests/push80/coverage_r7_push80f.rs
@@ -618,7 +618,7 @@ fn test_btree_delete_and_scan() {
     let mut current_root = root;
     for i in 1..=5 {
         let row = vec![Value::Integer(i), Value::Text(format!("row_{i}").into())];
-        current_root = btree.insert(current_root, i as i64, &row).unwrap();
+        current_root = btree.insert(current_root, i, &row).unwrap();
     }
 
     // Delete row 3
